Listar barcos com salário atrasado/Métodos Listar verificam se há o que listar

+++ codigo/Experiencia.java
@@ -16,7 +16,7 @@ public enum Experiencia {
         this.custoCurso = custoCurso;
     }
 
-    public double getSalario() {
+    public int getSalario() {
         return salario;
     }
